mariola/requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
astunparse==1.6.3
attrs==24.3.0
binance-python==0.1.3
certifi==2024.12.14
charset-normalizer==3.4.1
dateparser==1.2.0
flatbuffers==24.12.23
frozenlist==1.5.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.68.1
h11==0.14.0
h5py==3.12.1
haskellian==0.5.11
httpcore==1.0.7
httpx==0.28.1
idna==3.10
joblib==1.4.2
keras==3.7.0
lazy_loader==0.4
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.4.1
multidict==6.1.0
namex==0.0.8
numpy==1.26.4
opt_einsum==3.4.0
optree==0.13.1
orjson==3.10.12
packaging==24.2
pandas==2.2.3
propcache==0.2.1
protobuf==4.25.5
pycryptodome==3.21.0
pydantic==2.10.4
pydantic_core==2.27.2
Pygments==2.18.0
python-binance==1.0.26
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
ramda==0.7.6
regex==2024.11.6
requests==2.32.3
rich==13.9.4
scikit-learn==1.6.0
scipy==1.14.1
setuptools==75.6.0
six==1.17.0
sniffio==1.3.1
TA-Lib==0.4.32
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
termcolor==2.5.0
threadpoolctl==3.5.0
toolz==1.0.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.3.0
websockets==14.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.0
yarl==1.18.3
